+++ src/pages/Calendar.tsx
@@ -1,0 +1,9 @@
+
+
+export default function Calendar() {
+    return (
+        <>
+        <h1>This is going to be our calendar</h1>
+        </>
+    )
+}
